requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0